+++ App/Model/Comment.php
@@ -17,12 +17,12 @@ class Comment extends TableDataGateway
     /**
      * 主键字段
      */
-    public string $primaryKey = 'id';
+    public $primaryKey = 'id';
 
     /**
      * 定义从属关联：一个评论属于一个文章
      */
-    public array $belongsTo = [
+    public ?array $belongsTo = [
         [
             'tableClass' => Post::class,
             'foreignKey' => 'post_id',